Fixed small problems with local plugins. Clear should work now, too.
Adjusted plugins/experimental/template/CMakeLists.txt and
plugins/experimental/sample/CMakeLists.txt so that
they do not require explicit paths to the trunk or to the plugin sources. They
use cmake's internal variables now. This is the recommended way to write CMakeLists.txt
files for new plugins, since this allows to place a plugin anywhere on the filesystem
without adjusting the CMakeLists.txt.